Clean up refactor
