+++ packages/core/src/components/DndContext/index.ts
@@ -1,14 +1,8 @@
-export {
-  ActiveDraggableContext,
-  DndContext,
-  LayoutMeasuringStrategy,
-  LayoutMeasuringFrequency,
-} from './DndContext';
+export {ActiveDraggableContext, DndContext} from './DndContext';
 export type {
   CancelDrop,
   DragStartEvent,
   DragMoveEvent,
   DragOverEvent,
   DragEndEvent,
-  LayoutMeasuring,
 } from './DndContext';
