+++ packages/core/src/components/index.ts
@@ -3,18 +3,13 @@ export {
   defaultAnnouncements,
   ScreenReaderInstructions,
 } from './Accessibility';
-export {
-  DndContext,
-  LayoutMeasuringFrequency,
-  LayoutMeasuringStrategy,
-} from './DndContext';
+export {DndContext} from './DndContext';
 export type {
   CancelDrop,
   DragEndEvent,
   DragOverEvent,
   DragMoveEvent,
   DragStartEvent,
-  LayoutMeasuring,
 } from './DndContext';
 export {
   DragOverlay,
