+++ packages/core/src/hooks/index.ts
@@ -5,3 +5,5 @@ export {
 } from './useDraggable';
 export {useDndContext, UseDndContextReturnValue} from './useDndContext';
 export {useDroppable, UseDroppableArguments} from './useDroppable';
+export {LayoutMeasuringStrategy, LayoutMeasuringFrequency} from './utilities';
+export type {LayoutMeasuring} from './utilities';
